refactor(psxed-gltf): split 4.6k-line lib.rs into focused modules
Mechanical, behavior-preserving split of the glTF/FBX importer's monolithic
lib.rs (4,608 lines) into cohesive modules. lib.rs stays a ~1.1k-line facade
holding the public API and the shared cook data types.

  fbx.rs         FBX (ufbx) import backend: skinned mesh + animation cooking
  source.rs      glTF skinned-mesh source extraction (vertices/joints/matrices)
  model_blob.rs  cooks decimated source data into the runtime PSXM model blob
  animation.rs   glTF animation sampling + PSXM animation-clip cooking
  math.rs        leaf numeric helpers + little-endian blob writers
  tests.rs       the crate's unit tests

Shared cook types (Trs, SkinnedSourceMesh, ModelBounds, ...) stay in the crate
root so child modules reach them through `use super::*;`; moved items are
`pub(crate)`. No logic changes: all 16 tests pass, clippy is unchanged, and
psxed-project still builds against the identical public API.
